requirements.txt
flask
flask-login
pymongo
bcrypt
azure-cognitiveservices-vision-computervision
python-dotenv